make windows acceptance tests runnable locally

+++ ci/tasks/bbl-destroy.sh
@@ -3,40 +3,48 @@
 set -eu -o pipefail
 
 ROOT_DIR=$PWD
-REPO_DIR=$ROOT_DIR/envs-output
 
-git clone -q "file://$ROOT_DIR/envs" "$REPO_DIR"
+BOSH_BINARY_PATH=${BOSH_BINARY_PATH:-bosh}
+ROOT_DIR=${ROOT_DIR:-$PWD}
+
+REPO_DIR=$ROOT_DIR/envs
+
+if [ -z "${SKIP_GIT}" ]; then
+	REPO_DIR=$ROOT_DIR/envs-output
+
+	git clone -q "file://$ROOT_DIR/envs" "$REPO_DIR"
+
+	function commit_state {
+		cd "$REPO_DIR"
+
+		if [[ ! -n "$(git status --porcelain)" ]]; then
+			return
+		fi
+
+		git config user.name "${GIT_COMMITTER_NAME:-CI Bot}"
+		git config user.email "${GIT_COMMITTER_EMAIL:-ci@localhost}"
+		git add -A .
+		git commit -m "$ENV_NAME: bbl-destroy"
+	}
+
+	trap commit_state EXIT
+fi
 
 BBL_STATE_DIR=$REPO_DIR/$ENV_NAME
 
-cd "$BBL_STATE_DIR"
+source $BBL_STATE_DIR/.envrc
 
-function commit_state {
-  cd "$REPO_DIR"
+$BOSH_BINARY_PATH deployments | awk '{ print $1 }' | xargs -n1 $BOSH_BINARY_PATH -n delete-deployment -d
 
-  if [[ ! -n "$(git status --porcelain)" ]]; then
-    return
-  fi
+$BOSH_BINARY_PATH -n clean-up --all
 
-  git config user.name "${GIT_COMMITTER_NAME:-CI Bot}"
-  git config user.email "${GIT_COMMITTER_EMAIL:-ci@localhost}"
-  git add -A .
-  git commit -m "$ENV_NAME: bbl-destroy"
-}
-
-trap commit_state EXIT
-
-source $BBL_STATE_DIR/bosh.sh
-
-bosh deployments | awk '{ print $1 }' | xargs -I name bosh -n -d name delete-deployment
-
-bosh -n clean-up --all
-
-bosh -n delete-env $BBL_STATE_DIR/bosh-manifest.yml \
+$BOSH_BINARY_PATH -n delete-env $BBL_STATE_DIR/bosh-manifest.yml \
   --vars-store $BBL_STATE_DIR/creds.yml  \
   --state $BBL_STATE_DIR/state.json \
   -l <(bbl --state-dir=$BBL_STATE_DIR bosh-deployment-vars)
 
 bbl --state-dir=$BBL_STATE_DIR destroy --no-confirm --skip-if-missing
 
-git rm -rf .
+if [ -z "${SKIP_GIT}" ]; then
+	git rm -rf .
+fi
